pyproject.toml
[project]
name = "HCMUT-chatbot-ai-agent"
version = "0.1.0"
description = "HCMUT Chatbot AI Agent"
authors = [
    {name = "leeduy0403",email = "lhduy2003@gmail.com"},
    {name = "Tuong Van", email = "doannguyenduycuong@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.14"

dependencies = [
    "fastapi (==0.120.3)",
    "langgraph (==1.0.2)",
    "litellm (==1.79.0)",
    "loguru (==0.7.3)",
    "motor (==3.7.1)",
    "openai (==2.6.1)",
    "pydantic (==2.12.3)",
    "python-dotenv (==1.2.1)",
    "scikit-learn (==1.7.2)",
    "termcolor (==3.2.0)",
    "uvicorn (==0.38.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-huggingface (>=1.0.0,<2.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "pinecone (>=7.3.0,<8.0.0)",
    "unstructured[docx] (>=0.18.15,<0.19.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "joblib (>=1.5.2,<2.0.0)",
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
